Search now displays gitProjectState in chunks of 5
Search now stores gitProjectState as an array of arrays of 5
The response from the github api is sliced with the sliceChunk function
and stored in gitProjectState.

A conditional was added to the map function that renders the commit
components to prevent the page from trying to load them before the state is
updated. Right now it is hard coded to render the first page. In the future
it will render the index of the array that matches the active pagination button.